@@ -8,7 +8,7 @@ import 'cubit/weather_state.dart';
 import 'model/weather.dart';
 
 class HomePage extends StatefulWidget {
-  HomePage({Key key}) : super(key: key);
+  HomePage({Key? key}) : super(key: key);
 
   @override
   _HomePageState createState() => _HomePageState();
